+++ src/AreaChart/AreaChart.story.tsx
@@ -31,6 +31,7 @@ import { symbol, symbolStar } from 'd3-shape';
 import { Gradient, GradientStop } from '../common/Gradient';
 import { Stripes } from '../common/Mask';
 import { ChartDataShape } from '../common/data';
+import { schemes } from '../common/color';
 
 storiesOf('Charts/Area/Single Series', module)
   .add(
@@ -39,7 +40,7 @@ storiesOf('Charts/Area/Single Series', module)
       const height = number('Height', 250);
       const width = number('Width', 350);
       const lineStroke = number('Stroke Width', 4);
-      const fill = color('Color', '#418AD7');
+      const color = select('Color Scheme', schemes, 'cybertron');
       const interpolation = select(
         'Interpolation',
         {
@@ -59,7 +60,7 @@ storiesOf('Charts/Area/Single Series', module)
           series={
             <AreaSeries
               interpolation={interpolation}
-              colorScheme={[fill]}
+              colorScheme={color}
               line={<Line strokeWidth={lineStroke} />}
             />
           }
@@ -152,6 +153,7 @@ storiesOf('Charts/Area/Multi Series', module)
     () => {
       const height = number('Height', 350);
       const width = number('Width', 550);
+      const color = select('Color Scheme', schemes, 'cybertron');
       const data = object('Data', multiDateData);
 
       return (
@@ -159,14 +161,7 @@ storiesOf('Charts/Area/Multi Series', module)
           width={width}
           height={height}
           data={data}
-          series={
-            <AreaSeries
-              type="grouped"
-              colorScheme={chroma
-                .scale(['27efb5', '00bfff'])
-                .colors(data.length)}
-            />
-          }
+          series={<AreaSeries type="grouped" colorScheme={color} />}
         />
       );
     },
@@ -177,20 +172,14 @@ storiesOf('Charts/Area/Multi Series', module)
     () => {
       const height = number('Height', 350);
       const width = number('Width', 550);
+      const color = select('Color Scheme', schemes, 'cybertron');
       const data = object('Data', longMultiDateData);
 
       return (
         <AreaChart
           width={width}
           height={height}
-          series={
-            <AreaSeries
-              type="grouped"
-              colorScheme={chroma
-                .scale(['ACB7C9', '418AD7'])
-                .colors(data.length)}
-            />
-          }
+          series={<AreaSeries type="grouped" colorScheme={color} />}
           data={data}
         />
       );
@@ -216,19 +205,14 @@ storiesOf('Charts/Area/Multi Series', module)
     () => {
       const height = number('Height', 350);
       const width = number('Width', 550);
+      const color = select('Color Scheme', schemes, 'cybertron');
       const data = object('Data', multiDateData);
 
       return (
         <StackedAreaChart
           width={width}
           height={height}
-          series={
-            <StackedAreaSeries
-              colorScheme={chroma
-                .scale(['27efb5', '00bfff'])
-                .colors(data.length)}
-            />
-          }
+          series={<StackedAreaSeries colorScheme={color} />}
           data={data}
         />
       );
@@ -240,6 +224,7 @@ storiesOf('Charts/Area/Multi Series', module)
     () => {
       const height = number('Height', 350);
       const width = number('Width', 550);
+      const color = select('Color Scheme', schemes, 'cybertron');
       const data = object('Data', multiDateData);
 
       return (
@@ -247,13 +232,7 @@ storiesOf('Charts/Area/Multi Series', module)
           width={width}
           height={height}
           data={data}
-          series={
-            <StackedNormalizedAreaSeries
-              colorScheme={chroma
-                .scale(['27efb5', '00bfff'])
-                .colors(data.length)}
-            />
-          }
+          series={<StackedNormalizedAreaSeries colorScheme={color} />}
         />
       );
     },