+++ src/Sparkline/BarSparklineChart.tsx
@@ -7,6 +7,8 @@ import {
   LinearYAxisTickSeries,
   LinearXAxisTickSeries
 } from '../common/Axis';
+import { BarSeries } from '../BarChart';
+import { schemes } from '../common/color';
 
 export interface BarSparklineChartProps extends BarChartProps {
   data: ChartShallowDataShape[];
@@ -15,6 +17,7 @@ export interface BarSparklineChartProps extends BarChartProps {
 export class BarSparklineChart extends Component<BarSparklineChartProps> {
   static defaultProps: Partial<BarSparklineChartProps> = {
     gridlines: null,
+    series: <BarSeries colorScheme={schemes.cybertron[0]} />,
     yAxis: (
       <LinearYAxis
         type="value"