+++ src/common/index.ts
@@ -12,3 +12,4 @@ export * from './Tooltip';
 export * from './utils';
 export * from './ZoomPan';
 export * from './Motion';
+export * from './color';